requirements.txt
discord.py
requests
python-dotenv
google-generativeai>=0.7.0